// frontend/src/app/pages/authentication/data/store/authentication.enum.ts
export enum AuthActionsEnum {
  SIGN_IN = '[Auth] Sign in',
  SIGN_IN_SUCCESS = '[Auth] Sign in success',
  SIGN_IN_FAILED = '[Auth] Sign in failed',
  SIGN_UP = '[Auth] Sign up',
  SIGN_UP_SUCCESS = '[Auth] Sign up success',
  SIGN_UP_FAILED = '[Auth] Sign up failed',
  SIGN_OUT = '[Auth] Sign out',
  SIGN_OUT_SUCCESS = '[Auth] Sign out success',
  SIGN_OUT_FAILED = '[Auth] Sign out failed',
  TOGGLE_FORMS = '[Auth] Forms toggled',
  REFRESH_TOKEN = '[Auth] refresh token',
  REFRESH_TOKEN_SUCCESS = '[Auth] refresh success',
  REFRESH_TOKEN_FAILED = '[Auth] refresh failed',
  DO_NOTHING = '[Auth] Nothing',
}